requirements.txt
Flask
Flask_Cors
requests
pocket-api